code_refine
main 기준으로 정리
