+++ commands.py
@@ -1,0 +1,6 @@
+import random
+
+async def play_game(message):
+    choices = ['가위', '바위', '보']
+    bot_choice = random.choice(choices)
+    await message.channel.send(f'봇이 선택한 것은 {bot_choice}입니다!')
